+++ src/Core/Signal/Signal.hs
@@ -4,7 +4,7 @@ module Core.Signal.Signal where
 import Core.Character.Character
 
 
-data PlayerSignal = StrafeLeft | StrafeRight
+data PlayerSignal = StrafeLeft | StrafeRight | SwingLeft | SwingRight
                   deriving (Enum, Eq, Read, Show, Bounded, Ord)
 
 data FlowSignal = Progress
@@ -15,3 +15,5 @@ data Signal = PlayerSignal PlayerSignal | FlowSignal FlowSignal
 signalToSide :: PlayerSignal -> Side
 signalToSide StrafeLeft = Left'
 signalToSide StrafeRight = Right'
+signalToSide SwingLeft = Left'
+signalToSide SwingRight = Right'