+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Added
 
+- Enemies can be killed.
 - Enemies are added.
 
 # [1.0.0 - 2023-10-25]